fix: исправлены обработки ошибок

+++ backend/services/eventService.js
@@ -1,5 +1,6 @@
 const eventRepository = require("../repositories/eventRepository");
 const { validate: isUUID } = require("uuid");
+const userRepository = require("../repositories/userRepository");
 
 class EventService {
   async getAllEvents() {
@@ -17,39 +18,72 @@ class EventService {
     return event;
   }
 
-  async createEvent(eventData) {
+async createEvent(eventData) {
     const { title, description, date, location, createdBy } = eventData;
 
     if (!title || !date || !createdBy || !location) {
-      throw new Error("Все обязательные поля должны быть заполнены");
+        throw new Error("Все обязательные поля должны быть заполнены");
+    }
+
+    if (!isUUID(createdBy)) {
+        throw new Error("Некорректный UUID пользователя");
+    }
+
+    const userExists = await userRepository.getUserById(createdBy);
+
+    if (!userExists) {
+        throw new Error("Пользователь не найден");
     }
 
     return await eventRepository.createEvent({ title, description, date, location, createdBy });
-  }
+}
 
-  async updateEvent(id, eventData) {
-    const { title, date, createdBy, location } = eventData;
 
-    if (!title || !date || !createdBy || !location) {
-      throw new Error("Все обязательные поля должны быть заполнены");
+async updateEvent(id, eventData) {
+    if (!isUUID(id)) {
+        throw new Error("Некорректный ID");
     }
 
-    const updatedEvent = await eventRepository.updateEvent(id, eventData);
-    if (!updatedEvent) {
-      throw new Error("Мероприятие не найдено");
+    const event = await eventRepository.getEventById(id);
+    if (!event) {
+        throw new Error("Мероприятие не найдено");
     }
 
-    return updatedEvent;
-  }
+    const filteredData = Object.fromEntries(
+        Object.entries(eventData).filter(([_, value]) => value !== undefined)
+    );
+
+    if (Object.keys(filteredData).length === 0) {
+        throw new Error("Нет данных для обновления");
+    }
+
+    if (filteredData.createdBy) {
+        if (!isUUID(filteredData.createdBy)) {
+            throw new Error("Некорректный UUID пользователя");
+        }
+
+        const userExists = await userRepository.getUserById(filteredData.createdBy);
+        if (!userExists) {
+            throw new Error("Пользователь не найден");
+        }
+    }
+
+    return await eventRepository.updateEvent(id, filteredData);
+}
+
 
   async deleteEvent(id) {
+    if (!isUUID(id)) {
+        throw new Error("Некорректный ID");
+    }
+
     const deleted = await eventRepository.deleteEvent(id);
     if (!deleted) {
-      throw new Error("Мероприятие не найдено");
+        throw new Error("Мероприятие не найдено");
     }
 
     return { message: "Мероприятие удалено" };
-  }
+}
 }
 
 module.exports = new EventService();
